Add transaction csv and fixtures
