+++ spec/fixtures/bad/transactions.csv
@@ -1,0 +1,2 @@
+id,invoice_id,credit_card_number,credit_card_expiration_date,result,created_at,updated_at
+1,1,,,success,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
