+++ spec/fixtures/good/transactions.csv
@@ -1,0 +1,4 @@
+id,invoice_id,credit_card_number,credit_card_expiration_date,result,created_at,updated_at
+1,1,4654405418249632,,success,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+2,2,4580251236515201,,success,2012-03-27 14:54:09 UTC,2012-03-27 14:54:09 UTC
+3,4,4354495077693036,,success,2012-03-27 14:54:10 UTC,2012-03-27 14:54:10 UTC
